requirements.txt
aiogram==3.20.0.post0
yt-dlp==2025.4.30
imageio-ffmpeg==0.6.0
ytmusicapi==1.10.3
python-dotenv==1.1.0
requests==2.31.0 